+++ hutool-core/src/test/java/cn/hutool/core/codec/Base64Test.java
@@ -89,4 +89,8 @@ public class Base64Test {
 		final String s = Base64.decodeStr(result, "gbk");
 		Assert.assertEquals(orderDescription, s);
 	}
+
+	@Test
+	public void decodeEmojiTest(){
+	}
 }